# Conflicts:
#	app/src/main/java/com/sem4ikt/uni/recipefinderchatbot/fragment/FavoritesFragment.java

+++ app/src/test/java/com/sem4ikt/uni/recipefinderchatbot/model/spoonacular/NutrientModelTest.java
@@ -21,8 +21,8 @@ public class NutrientModelTest {
     @Test
     public void setCalories()
     {
-        nutrientModel.setCalories(1);
-        Assert.assertEquals((int)nutrientModel.getCalories(),1);
+        nutrientModel.setCalories(1.0);
+        Assert.assertEquals((double)nutrientModel.getCalories(),1);
     }
 
     @Test
